refactor: dto segregation, crash fix
