+++ src/auth/auth.service.ts
@@ -1,9 +1,9 @@
-import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import * as bcrypt from 'bcrypt';
 
-import { AuthPayloadDto } from './dto/auth.dto';
 import { UsersService } from 'src/users/users.service';
+import { RegisterPayloadDto, AuthPayloadDto } from './dto';
 
 @Injectable()
 export class AuthService {
@@ -29,7 +29,7 @@ export class AuthService {
         }
     }
 
-    async register(userData: AuthPayloadDto) {
+    async register(userData: RegisterPayloadDto) {
         const hashedPassword = await bcrypt.hash(userData.password, this.saltOrRounds);
         const user = await this.usersService.create({
             ...userData,
