+++ src/auth/dto/auth.dto.ts
@@ -1,14 +1,14 @@
-import { IsEmail, IsNotEmpty, IsOptional, IsString, Length } from 'class-validator';
+import { IsBoolean, IsEmail, IsNotEmpty, IsOptional, IsString } from 'class-validator';
 
 export class AuthPayloadDto {
-    @IsNotEmpty()
-    @Length(3)
-    name?: string;
     @IsEmail()
     email: string;
+  
     @IsNotEmpty()
-    password: string;
     @IsString()
+    password: string;
+  
+    @IsBoolean()
     @IsOptional()
-    avatar?: string;
+    rememberMe: boolean;
 }
