+++ src/auth/dto/index.ts
@@ -1,0 +1,2 @@
+export * from "./register.dto";
+export * from "./auth.dto";
