+++ src/users/users.controller.ts
@@ -19,7 +19,7 @@ export class UsersController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.usersService.findOne(id);
+    return this.usersService.getById(id);
   }
 
   @Patch(':id')
